switched potential function signature to double; added potential function compiler and evaluation functions

+++ src/ConsSeg_types.h
@@ -4,6 +4,6 @@
 #include <Rcpp.h>
 
 // signature of potential functions
-typedef long double (*funcPtr)(int, int);
+typedef long double (*funcPtr)(double, double);
 
 #endif
